feat: TableGroupDomainService 구
